+++ src/renderer/src/components/message/MessageBlock.tsx
@@ -9,6 +9,7 @@ import {
   MEMORY_TOOL_RETRIEVE,
   MEMORY_TOOL_DELETE,
   MEMORY_TOOL_LIST,
+  MEMORY_TOOL_UPDATE,
   POWER_TOOL_BASH,
   POWER_TOOL_FETCH,
   POWER_TOOL_FILE_EDIT,
@@ -55,6 +56,7 @@ import { StoreMemoryToolMessage } from './StoreMemoryToolMessage';
 import { RetrieveMemoryToolMessage } from './RetrieveMemoryToolMessage';
 import { DeleteMemoryToolMessage } from './DeleteMemoryToolMessage';
 import { ListMemoriesToolMessage } from './ListMemoriesToolMessage';
+import { UpdateMemoryToolMessage } from './UpdateMemoryToolMessage';
 import { ActivateSkillToolMessage } from './ActivateSkillToolMessage';
 import { areMessagesEqual } from './utils';
 
@@ -181,6 +183,8 @@ const MessageBlockComponent = ({ baseDir, taskId, message, allFiles, renderMarkd
             return <DeleteMemoryToolMessage message={toolMessage} onRemove={remove} compact={compact} />;
           case MEMORY_TOOL_LIST:
             return <ListMemoriesToolMessage message={toolMessage} onRemove={remove} compact={compact} />;
+          case MEMORY_TOOL_UPDATE:
+            return <UpdateMemoryToolMessage message={toolMessage} onRemove={remove} compact={compact} />;
           default:
             break;
         }